null safety migration

+++ lib/utils.dart
@@ -2,7 +2,7 @@ import 'dart:math';
 
 class Utils {
   static double fixed(double value, int decimal) {
-    int fac = pow(10, decimal);
+    int fac = pow(10, decimal) as int;
     return (value * fac).round() / fac;
   }
 }
